Create loops.py
For Loop:
A for loop repeats a block of code a set number of times. It’s commonly used when you know how many times you want to loop.

While Loop:
A while loop repeats as long as a condition is True. It's used when you don't know in advance how many times to loop.